This is in response to trac report # 19 which suggests a speedup for dijkstra().
The suggestion is to check for duplicates on the priority queue before 
adding a node.  I chose to have the algorithm check when removing
rather than when adding to the queue (more simple implementation).
This probably requires a bigger queue than checking while adding, but
it seems faster than searching the queue each time a node is added.
(? though maybe an efficient non-dup priority queue algorithm exists ?)

This fix at least removes the need to check for neighbors multiple times.
So it realizes the bulk of the speedup hinted at by trac # 19.



git-svn-id: http://networkx.lanl.gov/svn/networkx/trunk@67 3ed01bd8-26fb-0310-9e4c-ca1a4053419f