Add predecessor function.
def predecessor(G,source,target=False,cutoff=False):

Returns dictionary of predecessors for the path from source to all nodes in G.  

Optional target returns only predecessors between source and target.
Cutoff is a limit on the number of hops traversed.

Example for the path graph 0-1-2-3

>>> G=path_graph(4)
>>> print G.nodes()
[0, 1, 2, 3]
>>> predecessor(G,0)
{0: [], 1: [0], 2: [1], 3: [2]}

Fixes #26.

The code to build a shortest path from a predecessor dictionary
is left as an exercise.
 



git-svn-id: http://networkx.lanl.gov/svn/networkx/trunk@171 3ed01bd8-26fb-0310-9e4c-ca1a4053419f