In regard to discussions with Sverre Sundsdal I put in a second version of shortest_path_bi called shortest_path_bi2 for speed testing.  I will remove it when testing is done.
Also changed a bunch of stuff while working to understand
the code which I think ought to make it faster and perhaps 
also more readable-- though that is often a matter of taste.
I will revert any of these changes upon request.

I also lowercased variables in the older shortest_path
routines to match the naming standard we are using.



git-svn-id: http://networkx.lanl.gov/svn/networkx/trunk@244 3ed01bd8-26fb-0310-9e4c-ca1a4053419f