Add a method to process nbunch.  This could be used within base.py at the cost of a (local) function lookup and benefit of more readable code.  +1 from me, but won't do unless others agree.
Outside base.py and xbase.py, one can use this method
G.prepare_nbunch(nbunch)   to process the nbunch: make it
a list of nodes in nbunch which are also in G.
When we leave python 2.3 we can make this return an iterator
if desired.

I implemented this feature in clustering() and triangles().

tests for clustering all had answers of 0.0 or 1.0, so I added a twist.



git-svn-id: http://networkx.lanl.gov/svn/networkx/trunk@274 3ed01bd8-26fb-0310-9e4c-ca1a4053419f