Fixes #37.
Passing a digraph to a connected component function now
raises an exception.  This applies the terminology decsion that the 
term "connected" should be only be used for digraphs.

This solution may not be the final answer on implementation 
but should prevent some confusion for now.

It might be useful to implement reachability and strong connectiviy
for digraphs.



git-svn-id: http://networkx.lanl.gov/svn/networkx/trunk@286 3ed01bd8-26fb-0310-9e4c-ca1a4053419f