Changed code for degree method (and friends in_degree/out_degree) so that they call degree_iter and friends.  This allows the new result that the order of the degree values sent back in a list is the same as the order of the nodes in nbunch.  see ticket 75.
Added tests using a simple 3-node case when the order is not the
same using dict.values().

This is not finished work however because XDiGraph doesn't have 
degree_iter (and iter friends).  I think that means that XDiGraph's
degree_iter has been broken all along.  That was more than I wanted 
to mess with tonight.



git-svn-id: http://networkx.lanl.gov/svn/networkx/trunk@455 3ed01bd8-26fb-0310-9e4c-ca1a4053419f