Remove **kwds from __init__ on graph classes. We don't need it and it causes bugs (our tests for selfloops used the keyword loops=True and it wasn't caught).
Also cleaned up the __init__ methods defaults for name
and removed the self=   portion of self=convert.from_whatever()
because create_using does the work and self= creates a local
variable named 'self' that HAPPENS to point to the same structure as 
the argument variable 'self' and I thought that was prone to 
becoming buggy someday.

Added tests to check data= and name= for each of the four classes.



git-svn-id: http://networkx.lanl.gov/svn/networkx/trunk@496 3ed01bd8-26fb-0310-9e4c-ca1a4053419f