Add a predecessor routine for weighted networks called: dijkstra_predecessor_and_distance() to path.py. and put a hook in betweenness_centrality so that if you use the argument     weighted_edges=True betweenness is calculated using dijkstra shortest paths.
Also, removed many default arguments which were set to False
and then tested.  When the argument is set to 0 (a node name)
the code treated it as False (b/c False is 0).
So, I changed them to None and tested for None.

Added some tests to path and centrality modules but could use more.



git-svn-id: http://networkx.lanl.gov/svn/networkx/trunk@542 3ed01bd8-26fb-0310-9e4c-ca1a4053419f