Fixes #116
Change convert.from_dict_of_dicts so that it creates valid
networkx graphs when the dict_of_dicts is incomplete: that is
when the dict_of_dicts shows only one direction of an undirected edge.
This change also makes sure G.pred is updated for DiGraphs. 

I did not fix this issue for graphs with multiedges.

Also added code to make this routine work with XDiGraph when G.multiedge is True.



git-svn-id: http://networkx.lanl.gov/svn/networkx/trunk@613 3ed01bd8-26fb-0310-9e4c-ca1a4053419f