Fixed a logic error during update of GMState.inout_x and DiGMState.in_x, and DiGMState.out_x.
When updating GMState.inout_1 and GMState.inout_2, we need to check if a node already 
exists in the dictionary.  Otherwise, we will overwrite when the node was added to the 
dictionary---this screws up backtracking.

Also, minor documentation changes.

fixes #122

(this is a recommit so that the changeset is linked to closing of the ticket)

git-svn-id: http://networkx.lanl.gov/svn/networkx/trunk@680 3ed01bd8-26fb-0310-9e4c-ca1a4053419f