Speed up neighbors considerably by replacing return list(self.neighbors_iter(n))        with return self.adj[n].keys()      (with a try-except to catch errors)
This doesn't work for xgraph because current API returns a 
neighbor twice if there are multiple edges.  So XGraph goes
back to list(self.neighbor_iter(n)).

I think this is actually an incorrect definition of neighbors!
In neighbors, each neighbor should be listed once...
But I'm not going to change it in the trunk.  Look to NEP for that.  



git-svn-id: http://networkx.lanl.gov/svn/networkx/trunk@694 3ed01bd8-26fb-0310-9e4c-ca1a4053419f